+++ src/prelude.rs
@@ -1,4 +1,4 @@
 //! Prelude for crypto trait methods.
 
 pub use crate::secrets::Secret;
-pub use crate::signer::{JwsSigner, JwsSignerSync, SignedBytes};
+pub use crate::signer::{JwsSigner, JwsSignerSync};